--- grafana-plugin/src/containers/IntegrationForm/HookForm.module.scss
@@ -1,63 +1,0 @@
-.form {
-  width: 100%;
-}
-
-.extra-fields {
-  padding: 12px;
-  margin-bottom: 24px;
-  border: var(--border-weak);
-  border-radius: var(--border-radius);
-
-  &__radio {
-    margin-bottom: 12px;
-  }
-
-  &__icon {
-    margin-top: -4px;
-  }
-}
-
-.selectors-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: -15px;
-}
-
-.textarea:hover {
-  // TODO: change this to fetch from emotion instead
-}
-
-.collapse {
-  width: 100%;
-  margin-bottom: 24px;
-}
-
-.collapse svg {
-  color: var(--primary-text-link) !important;
-}
-
-.integration-info-list {
-  list-style-position: inside;
-  margin: 16px 0;
-}
-
-.integration-info-item {
-  margin-left: 16px;
-}
-
-.servicenow-heading {
-  margin-bottom: 16px;
-}
-
-.webhook-test {
-  margin-bottom: 16px;
-}
-
-.webhook-switch {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 24px;
-}
