+++ grafana-plugin/src/containers/IntegrationForm/IntegrationForm.styles.ts
@@ -1,0 +1,73 @@
+import { css } from '@emotion/css';
+import { GrafanaTheme2 } from '@grafana/data';
+
+export const getIntegrationFormStyles = (theme: GrafanaTheme2) => {
+  return {
+    form: css`
+      width: 100%;
+    `,
+
+    extraFields: css`
+      padding: 12px;
+      margin-bottom: 24px;
+      border: var(--border-weak);
+      border-radius: var(--border-radius);
+    `,
+
+    extraFieldsRadio: css`
+      margin-bottom: 12px;
+    `,
+
+    extraFieldsIcon: css`
+      margin-top: -4px;
+    `,
+
+    selectorsContainer: css`
+      width: 100%;
+      display: flex;
+      flex-direction: column;
+      margin-bottom: -15px;
+    `,
+
+    collapse: css`
+      width: 100%;
+      margin-bottom: 24px;
+
+      svg {
+        color: ${theme.colors.primary.text} !important;
+      }
+    `,
+
+    integrationInfoList: css`
+      list-style-position: inside;
+      margin: 16px 0;
+    `,
+
+    integrationInfoItem: css`
+      margin-left: 16px;
+    `,
+
+    serviceNowHeading: css`
+      margin-bottom: 16px;
+    `,
+
+    webhookTest: css`
+      margin-bottom: 16px;
+    `,
+
+    webhookSwitch: css`
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      gap: 8px;
+      margin-bottom: 24px;
+    `,
+
+    labels: css`
+      margin-bottom: 20px;
+    `,
+
+    // TODO: figure out grafana bug on border
+    textarea: css``,
+  };
+};
